pyproject.toml
[project]
name = "bioinfo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.4",
    "django-unfold>=0.90.0",
    "jinja2>=3.1.6",
    "pillow>=12.2.0",
    "sqlparse>=0.5.5",
]
